Corrected php namespace for tests and added php5.6 compatibility
