+++ tests/CompressionDeflateHandlerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sikei\React\Http\Middleware;
+namespace Sikei\React\Tests\Http\Middleware;
 
 use Clue\React\Zlib\ZlibFilterStream;
 use PHPUnit\Framework\TestCase;
@@ -8,11 +8,12 @@ use React\Http\HttpBodyStream;
 use React\Http\ServerRequest;
 use React\Stream\ThroughStream;
 use function RingCentral\Psr7\stream_for;
+use Sikei\React\Http\Middleware\CompressionDeflateHandler;
 
 class CompressionDeflateHandlerTest extends TestCase
 {
 
-    /** @var CompressionGzipHandler */
+    /** @var CompressionDeflateHandler */
     public $handler;
 
     public function setUp()
