+++ tests/CompressionGzipHandlerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sikei\React\Http\Middleware;
+namespace Sikei\React\Tests\Http\Middleware;
 
 use Clue\React\Zlib\ZlibFilterStream;
 use PHPUnit\Framework\TestCase;
@@ -8,6 +8,7 @@ use React\Http\HttpBodyStream;
 use React\Http\ServerRequest;
 use React\Stream\ThroughStream;
 use function RingCentral\Psr7\stream_for;
+use Sikei\React\Http\Middleware\CompressionGzipHandler;
 
 class CompressionGzipHandlerTest extends TestCase
 {
